att6300p_fdc.cpp: Use an explicitly sized type for m_rate, since it's registered for state saves.  Fix reading back of the rate control register.

+++ src/mame/olivetti/att6300p_fdc.h
@@ -26,5 +26,5 @@ protected:
     void rc_w(uint8_t data);
 
 private:
-    int m_rate;
+    uint8_t m_rate;
 };
